Refactor yum_helper.py for quicker transactions
Yum is a backed by a database and seems to crash and leave the database
in the middle of a transaction. This has resulted in what appears to be
a corrupt yum database.

In order to get around this usages of the yum db now open, do minimal
work, and then close. The idea here is to do many small quick
transactions rather than one massive transaction.