Add context manger to manage close of package db
The python file that is used to manage dnf and yum packages was written
for python 2.4 and sometimes had issues with leaving the package db
open. This was likely caused by the closing of the database happening in
__del__ method see
http://yum.baseurl.org/download/docs/yum-api/3.2.27/yum-pysrc.html#YumBase
on line 165. The problem with this method is that according to the
python 2.7 docs `It is not guaranteed that __del__() methods are called
for objects that still exist when the interpreter exits.` (see
https://docs.python.org/2.7/reference/datamodel.html#object.__del__  for
more info). The alternative was to use a python context manager and to
avoid using the package manager in both the global and local scopes
(which might have kept it away from the garbage collector).

There may have been a case were one or more package managers were
floating around  due to the use of the global scope. As a result, the
Wrappers are there to function as a place to hold all the context
information for a given run. These are currently not singletons, but
should not need to be as a yum/dnf lock would cause conflicts and
terminate a second run of the class, and there is only one generated in
the current code.

Indentation was also cleaned up as it seems that some ruby indentations
made it into this python class (read 2 space indents).

Also as python 2.4 no longer needs to be supported, some code was
removed to allow for better.

Signed-off-by: Brian Faherty <anothergenericuser@gmail.com>